requirements.txt
asgiref==3.4.1
astroid==2.4.2
backcall==0.2.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cirq==0.9.1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
freezegun==0.3.15
google-api-core==1.26.0
google-auth==1.27.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.1.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.4
mccabe==0.6.1
mpmath==1.2.1
networkx==2.5
numpy==1.20.1
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.2
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.16
protobuf==3.12.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.0
pylint==2.6.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
requests==2.25.1
rsa==4.7.1
scipy==1.6.1
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.2
sympy==1.7.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzdata==2021.5
urllib3==1.26.3
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.12.1
